+++ server-webapp/src/main/java/org/taverna/server/master/interfaces/SecurityContextFactory.java
@@ -7,7 +7,6 @@ package org.taverna.server.master.interfaces;
 
 import java.io.Serializable;
 
-import org.taverna.server.master.localworker.RemoteRunDelegate;
 import org.taverna.server.master.utils.UsernamePrincipal;
 
 /**
@@ -28,6 +27,6 @@ public interface SecurityContextFactory extends Serializable {
 	 * @throws Exception
 	 *             If anything goes wrong.
 	 */
-	TavernaSecurityContext create(RemoteRunDelegate run, UsernamePrincipal owner)
+	TavernaSecurityContext create(TavernaRun run, UsernamePrincipal owner)
 			throws Exception;
 }
