+++ server-webapp/src/main/java/org/taverna/server/master/worker/RunDatabaseDAO.java
@@ -3,9 +3,9 @@
  * 
  * See the file "LICENSE" for license terms.
  */
-package org.taverna.server.master.localworker;
+package org.taverna.server.master.worker;
 
-import static org.taverna.server.master.localworker.RunConnection.toDBform;
+import static org.taverna.server.master.worker.RunConnection.toDBform;
 
 import java.io.IOException;
 import java.util.ArrayList;
